make encoder the core binary serialization logic library
